% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_gday_params.R
\name{adjust_gday_params}
\alias{adjust_gday_params}
\title{adjust_gday_params}
\usage{
adjust_gday_params(in_fname, out_fname, replacements)
}
\arguments{
\item{in_fname}{location on disk where inputs are stored}

\item{out_fname}{prefix of input and output files}

\item{replacements}{location on disk where outputs will be stored}
}
\value{
generates GDAY .ini param file in the out_fname path
}
\description{
R Function to edit param file for GDAY
}
\author{
Martin De Kauwe
\dontrun{
in_fname <- "base_start.cfg"
out_fname <- "test.ini"
replacements <- list("ncycle" = "true",
                    "modeljm" = "3",
                    "print_options" = "end",
                    "jmax" = "110.0",
                    "vcmax" = "55.0")
adjust_gday_params(in_fname, out_fname, replacements)
}
}
