% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.GDAY.R
\name{model2netcdf.GDAY}
\alias{model2netcdf.GDAY}
\title{Function to convert GDAY model output to standard netCDF format}
\usage{
model2netcdf.GDAY(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of GDAY model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert GDAY output to netCDF
}
\details{
Converts all output contained in a folder to netCDF.
}
\author{
Martin De Kauwe
}
